+++ packages/simple_dart_code_gen/CHANGELOG.md
@@ -1,11 +1,16 @@
+## 0.2.0
+
+- fix toString() method output
+- support string literal interpolation
+
 ## 0.1.0
 
-support function parameter name
+- support function parameter name
 
 ## 0.0.2
 
-add output toString method
+- add output toString method
 
 ## 0.0.1
 
-initial release
+- initial release